+++ OpenSim/Moco/MocoCasADiSolver/CasOCProblem.cpp
@@ -33,10 +33,13 @@ namespace CasOC {
 
 Iterate Iterate::resample(const casadi::DM& newTimes,
                           bool appendProjectionStates = false) const {
-    auto mocoIt = OpenSim::convertToMocoTrajectory(*this);
+    // Since we are converting to a MocoTrajectory and immediately converting 
+    // back to a CasOC::Iterate after resampling, we do not need to provide 
+    // Input control indexes, even if they are present in the MocoProblem.
+    auto mocoTraj = OpenSim::convertToMocoTrajectory(*this);
     auto simtkNewTimes = OpenSim::convertToSimTKVector(newTimes);
-    mocoIt.resample(simtkNewTimes);
-    return OpenSim::convertToCasOCIterate(mocoIt, mocoIt.getSlackNames(),
+    mocoTraj.resample(simtkNewTimes);
+    return OpenSim::convertToCasOCIterate(mocoTraj, mocoIt.getSlackNames(),
             appendProjectionStates);
 }