+++ OpenSim/Moco/MocoCasADiSolver/MocoCasOCProblem.h
@@ -70,7 +70,8 @@ inline casadi::DM convertToCasADiDM(const SimTK::Vector& simtkVec) {
 /// This converts a MocoTrajectory to a CasOC::Iterate.
 inline CasOC::Iterate convertToCasOCIterate(const MocoTrajectory& mocoTraj,
         const std::vector<std::string>& expectedSlackNames,
-        bool appendProjectionStates = false) {
+        bool appendProjectionStates = false,
+        std::vector<int> inputControlIndexes = {}) {
     CasOC::Iterate casIt;
     CasOC::VariablesDM& casVars = casIt.variables;
     using CasOC::Var;
@@ -78,8 +79,36 @@ inline CasOC::Iterate convertToCasOCIterate(const MocoTrajectory& mocoTraj,
     casVars[Var::final_time] = mocoTraj.getFinalTime();
     casVars[Var::states] =
             convertToCasADiDMTranspose(mocoTraj.getStatesTrajectory());
-    casVars[Var::controls] =
+
+    casadi::DM controls = 
             convertToCasADiDMTranspose(mocoTraj.getControlsTrajectory());
+    casadi::DM input_controls = 
+            convertToCasADiDMTranspose(mocoTraj.getInputControlsTrajectory());
+    std::vector<std::string> controlNames = mocoTraj.getControlNames();
+    std::vector<std::string> inputControlNames = mocoTraj.getInputControlNames();
+    std::vector<std::string> casControlNames;
+    int numTotalControls = static_cast<int>(controls.rows()) + 
+            static_cast<int>(input_controls.rows());
+    casadi::DM casControls(numTotalControls, controls.columns());
+    
+    int ic = 0;
+    int iic = 0;
+    std::sort(inputControlIndexes.begin(), inputControlIndexes.end());
+    int numInputControls = static_cast<int>(inputControlIndexes.size());
+    for (int i = 0; i < numTotalControls; ++i) {
+        if (iic < numInputControls && inputControlIndexes[iic] == i) {
+            casControls(i, casadi::Slice()) = 
+                    input_controls(iic, casadi::Slice());
+            casControlNames.push_back(inputControlNames[iic]);
+            ++iic;
+        } else {
+            casControls(i, casadi::Slice()) = controls(ic, casadi::Slice());
+            casControlNames.push_back(controlNames[ic]);
+            ++ic;
+        }
+    }
+    casVars[Var::controls] = casControls;
+            
     casVars[Var::multipliers] =
             convertToCasADiDMTranspose(mocoTraj.getMultipliersTrajectory());
 
@@ -95,7 +124,7 @@ inline CasOC::Iterate convertToCasOCIterate(const MocoTrajectory& mocoTraj,
     }
     casIt.times = convertToCasADiDMTranspose(mocoTraj.getTime());
     casIt.state_names = mocoTraj.getStateNames();
-    casIt.control_names = mocoTraj.getControlNames();
+    casIt.control_names = casControlNames;
     casIt.multiplier_names = mocoTraj.getMultiplierNames();
     casIt.derivative_names = mocoTraj.getDerivativeNames();
     casIt.parameter_names = mocoTraj.getParameterNames();
@@ -182,16 +211,41 @@ inline SimTK::Matrix convertToSimTKMatrix(const casadi::DM& casMatrix) {
 }
 
 template <typename TOut = MocoTrajectory>
-TOut convertToMocoTrajectory(const CasOC::Iterate& casIt) {
+TOut convertToMocoTrajectory(const CasOC::Iterate& casIt, 
+        std::vector<int> inputControlIndexes = {}) {
     SimTK::Matrix simtkStates;
     const auto& casVars = casIt.variables;
     using CasOC::Var;
     if (!casIt.state_names.empty()) {
         simtkStates = convertToSimTKMatrix(casVars.at(Var::states));
     }
+    int numTotalControls = static_cast<int>(casIt.control_names.size());
+    int numInputControls = static_cast<int>(inputControlIndexes.size());
+    int numControls = numTotalControls - numInputControls;
+    OPENSIM_ASSERT(numControls >= 0);
     SimTK::Matrix simtkControls;
-    if (!casIt.control_names.empty()) {
-        simtkControls = convertToSimTKMatrix(casVars.at(Var::controls));
+    SimTK::Matrix simtkInputControls;
+    std::vector<std::string> controlNames;
+    std::vector<std::string> inputControlNames;
+    if (numTotalControls) {
+        SimTK::Matrix allControls = 
+                convertToSimTKMatrix(casVars.at(Var::controls));
+        simtkControls.resize(allControls.nrow(), numControls);
+        simtkInputControls.resize(allControls.nrow(), numInputControls);
+        int ic = 0;
+        int iic = 0;
+        std::sort(inputControlIndexes.begin(), inputControlIndexes.end());
+        for (int i = 0; i < allControls.ncol(); ++i) {
+            if (iic < numInputControls && inputControlIndexes[iic] == i) {
+                simtkInputControls.updCol(iic) = allControls.col(i);
+                inputControlNames.push_back(casIt.control_names[i]);
+                ++iic;
+            } else {
+                simtkControls.updCol(ic) = allControls.col(i);
+                controlNames.push_back(casIt.control_names[i]);
+                ++ic;
+            }
+        }
     }
     SimTK::Matrix simtkMultipliers;
     if (!casIt.multiplier_names.empty()) {
@@ -219,9 +273,10 @@ TOut convertToMocoTrajectory(const CasOC::Iterate& casIt) {
     }
     SimTK::Vector simtkTimes = convertToSimTKVector(casIt.times);
 
-    TOut mocoTraj(simtkTimes, casIt.state_names, casIt.control_names,
-            casIt.multiplier_names, derivativeNames, casIt.parameter_names,
-            simtkStates, simtkControls, simtkMultipliers, simtkDerivatives,
+    TOut mocoTraj(simtkTimes, casIt.state_names, controlNames, 
+            inputControlNames, casIt.multiplier_names, derivativeNames, 
+            casIt.parameter_names, simtkStates, simtkControls, 
+            simtkInputControls, simtkMultipliers, simtkDerivatives, 
             simtkParameters);
 
     // Append slack variables. MocoTrajectory requires the slack variables to be
@@ -660,6 +715,10 @@ private:
             for (int ic = 0; ic < getNumControls(); ++ic) {
                 simtkControls[ic] = *(controls.ptr() + ic);
             }
+            // Updating the Inputs to InputControllers via the 
+            // ControlDistributor does not mark the model controls cache as 
+            // invalid, so we must do it manually here.
+            model.markControlsAsInvalid(simtkState);
         }
     }
     /// Apply variables from the optimizer to the MocoProblemRep's model and