+++ OpenSim/Moco/MocoGoal/MocoPeriodicityGoal.h
@@ -54,15 +54,12 @@ variable indicates the final state/control variable. Set the initial and
 final variables to the same variable to enforce periodicity on a single
 continuous variable. Only state and control continuous variable pairs are
 supported, and are specified via the 'state_pairs' and 'control_pairs'
-properties.
+properties. Control pairs can include both controls (i.e., model controls) and 
+Input controls.
 
 To handle initial and final variable values that are equal in absolute value
 but differ in sign (e.g. a pelvis rotation in walking), use
 addNegatedStatePair or addNegatedControlPair.
-
-If you wish to constrain all control signal pairs except those containing a
-control associated with a user-defined controller (e.g., PrescribedController),
-pass 'true' to `setIgnoreControlledActuators()`.
 
 To impose bilateral symmetry in a walking simulation,
 we can simulate over half a gait cycle and impose periodic constraints. For
@@ -125,16 +122,6 @@ public:
         append_control_pairs(std::move(pair));
     }
 
-    /// If true, do not constrain controls associated with user-defined
-    /// controllers.
-    void setIgnoreControlledActuators(bool v) {
-        set_ignore_controlled_actuators(v);
-    }
-    /// @copydoc setIgnoreControlledActuators()
-    bool getIgnoreControlledActuators() const {
-        return get_ignore_controlled_actuators();
-    }
-
 protected:
     bool getSupportsEndpointConstraintImpl() const override { return true; }
     Mode getDefaultModeImpl() const override {
@@ -150,12 +137,10 @@ private:
             "Periodic pairs of states.");
     OpenSim_DECLARE_LIST_PROPERTY(control_pairs, MocoPeriodicityGoalPair,
             "Periodic pairs of controls.");
-    OpenSim_DECLARE_PROPERTY(ignore_controlled_actuators, bool,
-            "If true, do not constrain controls belonging to actuators "
-            "controlled by user-defined controllers (default: false).");
     void constructProperties();
     mutable std::vector<std::tuple<int, int, int>> m_indices_states;
     mutable std::vector<std::tuple<int, int, int>> m_indices_controls;
+    mutable std::vector<std::pair<bool, bool>> m_isInputControl;
     mutable std::vector<std::pair<std::string,std::string>> m_state_names;
     mutable std::vector<std::pair<std::string,std::string>> m_control_names;
 };